task07 atm implementation

+++ task07/src/main/java/rem/hw07/atm/MoneyPar.java
@@ -16,7 +16,7 @@ public enum MoneyPar {
         this.nominal = nominal;
     }
 
-    public long getNominal() {
+    public Integer getNominal() {
         return nominal;
     }
 }
